Testing: Refactor testing framework cilicon
Testing, Cilicon: Add result getter and detailed result print function

Testing: Print total result at the end and exit program depending on all
tests passed or not

Testing, Cilicon: Add Context to subgroup tests. For example do the
cursor tests have valid and invalid movement tests.
