+++ test/buffer_test.c
@@ -1,4 +1,4 @@
-#include "buffer_test.h"
+#include "femto_test.h"
 
 void test_create_buffer()
 {
