--- test/buffer_test.h
@@ -1,12 +1,0 @@
-#ifndef BUFFER_TEST_H
-#define BUFFER_TEST_H
-
-#include "test_util.h"
-#include <buffer.h>
-
-void test_create_buffer();
-void test_append_to_buffer();
-void test_append_to_buffer_multiple_times();
-void test_suite_buffer();
-
-#endif
