--- test/cursor_test.h
@@ -1,8 +1,0 @@
-#ifndef CURSOR_TEST_H
-#define CURSOR_TEST_H
-
-#include "test_util.h"
-#include <session.h>
-
-void test_suite_cursor();
-#endif
