+++ test/session_test.c
@@ -1,4 +1,4 @@
-#include "session_test.h"
+#include "femto_test.h"
 
 void test_session_init()
 {
