--- test/session_test.h
@@ -1,13 +1,0 @@
-#ifndef SESSION_TEST_H
-#define SESSION_TEST_H
-
-#include "test_util.h"
-#include <session.h>
-#include <buffer.h>
-
-void test_load_file();
-void test_load_file_with_long_lines();
-void test_save_file();
-void test_suite_session();
-
-#endif
