+++ test/util_test.c
@@ -1,4 +1,4 @@
-#include "util_test.h"
+#include "femto_test.h"
 
 void test_clamp()
 {
