--- test/util_test.h
@@ -1,10 +1,0 @@
-#ifndef UTIL_TEST_C
-#define UTIL_TEST_C
-
-#include "test_util.h"
-#include <helper.h>
-
-void test_clamp();
-void test_suite_util();
-
-#endif // UTIL_TEST_C
